# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Core service broker with REST endpoints",
        "title": "Core API",
        "contact": {
            "email": "sercan@teamlify.co"
        },
        "version": "1.0"
    },
    "basePath": "/core/v1",
    "paths": {
...
        "/common/filters/sell-rep": {
            "get": {
                "description": "Func lists seller rep",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Common"
                ],
                "summary": "Func list sell rep by jwt token org_id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "search",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.HandlerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/entities.HandlerResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/entities.HandlerResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entities.HandlerResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    }
}